re-organize utilities and types
