+++ src/utilities/oas-schema/index.ts
@@ -1,0 +1,1 @@
+export { default } from './oas-schema';
