+++ src/utilities/oas-validator/index.ts
@@ -1,0 +1,1 @@
+export { default } from './oas-validator';
