+++ src/utilities/parameter-wrapper/index.ts
@@ -1,0 +1,1 @@
+export { default } from './parameter-wrapper';
